requirements.txt
pyaudio
requests
numpy
simplecoremidi
flask
python-dotenv
Pillow
ndicapi
imutils
lib
receiver
finder
